Add test for polyline generation
