+++ src/main/java/org/opentripplanner/util/PolylineEncoder.java
@@ -4,6 +4,7 @@ import java.util.AbstractList;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.locationtech.jts.geom.Point;
 import org.locationtech.jts.geom.Polygon;
 import org.opentripplanner.util.model.EncodedPolylineBean;
 
@@ -42,6 +43,9 @@ public class PolylineEncoder {
         } else if (geometry instanceof Polygon) {
             Polygon polygon = (Polygon) geometry;
             return createEncodings(new CoordinateList(polygon.getCoordinates()));
+        } else if (geometry instanceof Point) {
+            Point point = (Point) geometry;
+            return createEncodings(new CoordinateList(point.getCoordinates()));
         } else {
             throw new IllegalArgumentException(geometry.toString());
         }
